requirements.txt
pyTelegramBotAPI
requests
python-dotenv
psycopg2-binary
